Refactor(helm): extract type declarations into types.go
